findline tune

+++ src/lib/lfindline_char.cxx
@@ -376,7 +376,7 @@ TESTS:
 
     dx = abs(x1 - x2);
     dy = abs(y1 - y2);
-    dist= sqrt(dx*dx+dy*dy);
+    dist= sqrt(double(dx*dx+dy*dy));
 
     incx = 1;
     if (x2 < x1) incx = -1;
